+++ res/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Scripting-setting to select whether to upload the editor content or file content.
 * Autocompletion and bracket matching completion improvements.
 * Support for selecting multiple files when saving and deleting files from the log browser.
+* Removed firmware and package tabs and replaced them with buttons.
 
 ### 6.02
 #### Released 2023-03-12